Refactoring how we set the default parameter values
Moving the logic into `MeasureEvaluationSeeder`, where an already existing `ModelResolver` is ready for use.
