+++ cohort-engine/src/main/java/com/ibm/cohort/engine/measure/seed/MeasureEvaluationSeeder.java
@@ -3,11 +3,15 @@ package com.ibm.cohort.engine.measure.seed;
 import java.util.Date;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 
+import com.ibm.cohort.engine.measure.parameter.UnsupportedFhirTypeException;
 import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.lang3.tuple.Triple;
+import org.hl7.fhir.r4.model.Extension;
 import org.hl7.fhir.r4.model.Library;
 import org.hl7.fhir.r4.model.Measure;
+import org.hl7.fhir.r4.model.Type;
 import org.opencds.cqf.common.helpers.DateHelper;
 import org.opencds.cqf.common.helpers.UsingHelper;
 import org.opencds.cqf.common.providers.LibraryResolutionProvider;
@@ -15,6 +19,7 @@ import org.opencds.cqf.cql.engine.data.DataProvider;
 import org.opencds.cqf.cql.engine.debug.DebugMap;
 import org.opencds.cqf.cql.engine.execution.Context;
 import org.opencds.cqf.cql.engine.execution.LibraryLoader;
+import org.opencds.cqf.cql.engine.model.ModelResolver;
 import org.opencds.cqf.cql.engine.runtime.DateTime;
 import org.opencds.cqf.cql.engine.runtime.Interval;
 import org.opencds.cqf.cql.engine.terminology.TerminologyProvider;
@@ -23,6 +28,9 @@ import com.google.common.annotations.VisibleForTesting;
 import com.ibm.cohort.engine.measure.LibraryHelper;
 
 public class MeasureEvaluationSeeder {
+
+	// TODO: Tests need this...tests not in this class
+	public static final String PARAMETER_EXTENSION_URL = "http://hl7.org/fhir/us/cqfmeasures/StructureDefinition/cqfm-parameter";
 
 	private final TerminologyProvider terminologyProvider;
 	private final Map<String, DataProvider> dataProviders;
@@ -77,6 +85,14 @@ public class MeasureEvaluationSeeder {
 		Interval measurementPeriod = createMeasurePeriod(periodStart, periodEnd);
 		Context context = createContext(primaryLibrary, lastModelUri, dataProvider, measurementPeriod, productLine);
 
+		measure.getExtension().stream()
+				.filter(this::isParameterExtension)
+				.forEach(measureDefault ->
+						context.setParameter(
+								null,
+								measureDefault.getId(),
+								toCqlObject(measureDefault.getValue(), dataProvider)));
+
 		return new CustomMeasureEvaluationSeed(measure, context, measurementPeriod, dataProvider);
 	}
 
@@ -127,4 +143,13 @@ public class MeasureEvaluationSeeder {
 							DateHelper.resolveRequestDate(periodEnd, false), true);
 	}
 
+	private Object toCqlObject(Type type, ModelResolver modelResolver) {
+		return Optional.ofNullable(type)
+				.map((fhirType) -> modelResolver.resolvePath(fhirType, "value"))
+				.orElseThrow(() -> new UnsupportedFhirTypeException(type));
+	}
+
+	private boolean isParameterExtension(Extension extension) {
+		return PARAMETER_EXTENSION_URL.equals(extension.getUrl());
+	}
 }
