+++ cohort-engine/src/test/java/com/ibm/cohort/engine/measure/MeasureEvaluatorTest.java
@@ -21,6 +21,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import com.ibm.cohort.engine.measure.seed.MeasureEvaluationSeeder;
 import org.hl7.fhir.r4.model.Address;
 import org.hl7.fhir.r4.model.CanonicalType;
 import org.hl7.fhir.r4.model.CapabilityStatement;
@@ -385,7 +386,7 @@ public class MeasureEvaluatorTest extends BaseMeasureTest {
 
 		Measure measure = getProportionMeasure("ProportionMeasureName", library, expressionsByPopulationType);
 		Extension parameterExtension = new Extension();
-		parameterExtension.setUrl(MeasureEvaluator.PARAMETER_EXTENSION_URL);
+		parameterExtension.setUrl(MeasureEvaluationSeeder.PARAMETER_EXTENSION_URL);
 		parameterExtension.setId("SomeAge");
 
 		Type age = new IntegerType(20);
@@ -418,7 +419,7 @@ public class MeasureEvaluatorTest extends BaseMeasureTest {
 		Measure measure = getProportionMeasure("ProportionMeasureName", library, expressionsByPopulationType);
 
 		Extension parameterExtension = new Extension();
-		parameterExtension.setUrl(MeasureEvaluator.PARAMETER_EXTENSION_URL);
+		parameterExtension.setUrl(MeasureEvaluationSeeder.PARAMETER_EXTENSION_URL);
 		parameterExtension.setId("SomeAge");
 		Type age = new StringType("invalid");
 
@@ -448,7 +449,7 @@ public class MeasureEvaluatorTest extends BaseMeasureTest {
 		Measure measure = getProportionMeasure("ProportionMeasureName", library, expressionsByPopulationType);
 
 		Extension parameterExtension = new Extension();
-		parameterExtension.setUrl(MeasureEvaluator.PARAMETER_EXTENSION_URL);
+		parameterExtension.setUrl(MeasureEvaluationSeeder.PARAMETER_EXTENSION_URL);
 		parameterExtension.setId("SomeAge");
 
 		Address unsupportedType = new Address();
